docs(accordion): update accordion docs and structure
