+++ .storybook/main.js
@@ -1,7 +1,8 @@
 module.exports = {
   stories: [
     '../projects/canopy/src/lib/intro.stories.mdx',
-    '../projects/canopy/src/lib/accordion/accordion.stories.ts',
+    '../projects/canopy/src/lib/accordion/docs/guide.stories.mdx',
+    '../projects/canopy/src/lib/accordion/docs/accordion.stories.ts',
     '../projects/canopy/src/lib/alert/alert.stories.ts',
     '../projects/canopy/src/lib/brand-icon/brand-icons.stories.ts',
     '../projects/canopy/src/lib/breadcrumb/docs/guide.stories.mdx',
